+++ packages/stream_chat_flutter/lib/src/channel_bottom_sheet.dart
@@ -7,7 +7,7 @@ import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 @Deprecated("Use 'StreamChannelInfoBottomSheet' instead")
 class ChannelBottomSheet extends StatefulWidget {
   /// Constructor for creating bottom sheet
-  const ChannelBottomSheet({Key? key, this.onViewInfoTap}) : super(key: key);
+  const ChannelBottomSheet({super.key, this.onViewInfoTap});
 
   /// Callback when 'View Info' is tapped
   final VoidCallback? onViewInfoTap;
@@ -33,7 +33,6 @@ class _ChannelBottomSheetState extends State<ChannelBottomSheet> {
 
     final userAsMember = members
         .firstWhere((e) => e.user?.id == _streamChatState.currentUser?.id);
-    final isOwner = userAsMember.role == 'owner';
 
     return Material(
       color: _streamChatThemeData.colorTheme.barsBg,
@@ -179,9 +178,8 @@ class _ChannelBottomSheetState extends State<ChannelBottomSheet> {
                       });
                     },
                   ),
-                if (isOwner &&
-                    channel.ownCapabilities
-                        .contains(PermissionType.deleteChannel))
+                if (channel.ownCapabilities
+                    .contains(PermissionType.deleteChannel))
                   StreamOptionListTile(
                     leading: Padding(
                       padding: const EdgeInsets.symmetric(horizontal: 16),