+++ packages/stream_chat_flutter/lib/src/message_widget.dart
@@ -52,7 +52,7 @@ typedef MessageWidget = StreamMessageWidget;
 class StreamMessageWidget extends StatefulWidget {
   /// Creates a new instance of the message widget.
   StreamMessageWidget({
-    Key? key,
+    super.key,
     required this.message,
     required this.messageTheme,
     this.reverse = false,
@@ -105,7 +105,7 @@ class StreamMessageWidget extends StatefulWidget {
     this.customActions = const [],
     this.onAttachmentTap,
     this.usernameBuilder,
-  })  : attachmentBuilders = {
+  }) : attachmentBuilders = {
           'image': (context, message, attachments) {
             final border = RoundedRectangleBorder(
               borderRadius: attachmentBorderRadiusGeometry ?? BorderRadius.zero,
@@ -261,8 +261,7 @@ class StreamMessageWidget extends StatefulWidget {
                   .toList(),
             );
           },
-        }..addAll(customAttachmentBuilders ?? {}),
-        super(key: key);
+        }..addAll(customAttachmentBuilders ?? {});
 
   /// Function called on mention tap
   final void Function(User)? onMentionTap;
@@ -878,7 +877,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
           const Offstage();
     }
 
-    final children = <Widget>[];
+    final children = <WidgetSpan>[];
 
     final threadParticipants = widget.message.threadParticipants?.take(2);
     final showThreadParticipants = threadParticipants?.isNotEmpty == true;
@@ -909,36 +908,27 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
     const usernameKey = Key('username');
 
     children.addAll([
-      if (showInChannel || showThreadReplyIndicator) ...[
-        if (showThreadParticipants)
-          SizedBox.fromSize(
-            size: Size((threadParticipants!.length * 8.0) + 8, 16),
-            child: _buildThreadParticipantsIndicator(threadParticipants),
-          ),
-        InkWell(
-          onTap: widget.onThreadTap != null ? onThreadTap : null,
-          child: Text(msg, style: widget.messageTheme.repliesStyle),
-        ),
-      ],
-      if (showUsername) _buildUsername(usernameKey),
+      if (showUsername) WidgetSpan(child: _buildUsername(usernameKey)),
       if (showTimeStamp)
-        Text(
-          Jiffy(widget.message.createdAt.toLocal()).jm,
-          style: widget.messageTheme.createdAtStyle,
+        WidgetSpan(
+          child: Text(
+            Jiffy(widget.message.createdAt.toLocal()).jm,
+            style: widget.messageTheme.createdAtStyle,
+          ),
         ),
-      if (showSendingIndicator) _buildSendingIndicator(),
+      if (showSendingIndicator)
+        WidgetSpan(
+          child: _buildSendingIndicator(),
+        ),
     ]);
 
     final showThreadTail = !(hasUrlAttachments || isGiphy || isOnlyEmoji) &&
         (showThreadReplyIndicator || showInChannel);
 
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.end,
-      mainAxisAlignment:
-          widget.reverse ? MainAxisAlignment.end : MainAxisAlignment.start,
-      children: [
-        if (showThreadTail && !widget.reverse)
-          Container(
+    final threadIndicatorWidgets = <WidgetSpan>[
+      if (showThreadTail)
+        WidgetSpan(
+          child: Container(
             margin: EdgeInsets.only(
               bottom: context.textScaleFactor *
                   ((widget.messageTheme.repliesStyle?.fontSize ?? 1) / 2),
@@ -952,34 +942,38 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
               ),
             ),
           ),
-        ...children.map(
-          (child) {
-            Widget mappedChild = SizedBox(
-              height: context.textScaleFactor * 14,
-              child: child,
-            );
-            if (child.key == usernameKey) {
-              mappedChild = Flexible(child: mappedChild);
-            }
-            return mappedChild;
-          },
         ),
-        if (showThreadTail && widget.reverse)
-          Container(
-            margin: EdgeInsets.only(
-              bottom: context.textScaleFactor *
-                  ((widget.messageTheme.repliesStyle?.fontSize ?? 1) / 2),
-            ),
-            child: CustomPaint(
-              size: const Size(16, 32) * context.textScaleFactor,
-              painter: _ThreadReplyPainter(
-                context: context,
-                color: widget.messageTheme.messageBorderColor,
-                reverse: widget.reverse,
-              ),
+      if (showInChannel || showThreadReplyIndicator) ...[
+        if (showThreadParticipants)
+          WidgetSpan(
+            child: SizedBox.fromSize(
+              size: Size((threadParticipants!.length * 8.0) + 8, 16),
+              child: _buildThreadParticipantsIndicator(threadParticipants),
             ),
           ),
-      ].insertBetween(const SizedBox(width: 8)),
+        WidgetSpan(
+          child: InkWell(
+            onTap: widget.onThreadTap != null ? onThreadTap : null,
+            child: Text(msg, style: widget.messageTheme.repliesStyle),
+          ),
+        ),
+      ],
+    ];
+
+    if (widget.reverse) {
+      children.addAll(threadIndicatorWidgets.reversed);
+    } else {
+      children.insertAll(0, threadIndicatorWidgets);
+    }
+
+    return Text.rich(
+      TextSpan(
+        children: [
+          ...children,
+        ].insertBetween(const WidgetSpan(child: SizedBox(width: 8))),
+      ),
+      maxLines: 1,
+      textAlign: widget.reverse ? TextAlign.right : TextAlign.left,
     );
   }
 
@@ -1085,7 +1079,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
             showTimestamp: false,
             translateUserAvatar: false,
             showSendingIndicator: false,
-            padding: const EdgeInsets.all(0),
+            padding: EdgeInsets.zero,
             showReactionPickerIndicator: widget.showReactions &&
                 (widget.message.status == MessageSendingStatus.sent) &&
                 channel.ownCapabilities.contains(PermissionType.sendReaction),
@@ -1116,6 +1110,13 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
               widget.onThreadTap != null,
           showFlagButton: widget.showFlagButton,
           customActions: widget.customActions,
+          showDeleteMessage: widget.showDeleteMessage || isDeleteFailed,
+          showEditMessage: widget.showEditMessage &&
+              !isDeleteFailed &&
+              !widget.message.attachments
+                  .any((element) => element.type == 'giphy'),
+          showPinButton: widget.showPinButton,
+          showReactions: widget.showReactions,
         ),
       ),
     );
@@ -1142,7 +1143,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
             showTimestamp: false,
             translateUserAvatar: false,
             showSendingIndicator: false,
-            padding: const EdgeInsets.all(0),
+            padding: EdgeInsets.zero,
             showReactionPickerIndicator: widget.showReactions &&
                 (widget.message.status == MessageSendingStatus.sent) &&
                 channel.ownCapabilities.contains(PermissionType.sendReaction),
@@ -1268,6 +1269,7 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
         );
         if (isMessageRead) {
           child = Row(
+            mainAxisSize: MainAxisSize.min,
             children: [
               if (memberCount > 2)
                 Text(
@@ -1405,11 +1407,9 @@ class _StreamMessageWidgetState extends State<StreamMessageWidget>
 
 class _ThreadParticipants extends StatelessWidget {
   const _ThreadParticipants({
-    Key? key,
     required StreamChatThemeData streamChatTheme,
     required this.threadParticipants,
-  })  : _streamChatTheme = streamChatTheme,
-        super(key: key);
+  }) : _streamChatTheme = streamChatTheme;
 
   final StreamChatThemeData _streamChatTheme;
   final Iterable<User> threadParticipants;
@@ -1432,7 +1432,7 @@ class _ThreadParticipants extends StatelessWidget {
             padding: const EdgeInsets.all(1),
             child: StreamUserAvatar(
               user: user,
-              constraints: BoxConstraints.loose(const Size.fromRadius(7)),
+              constraints: BoxConstraints.tight(const Size.fromRadius(7)),
               showOnlineStatus: false,
             ),
           ),